refactor: move code to open project picker to new command
