+++ test/disabled/installGuide.test.ts
@@ -41,7 +41,8 @@ describe('Install guide', () => {
       mockSingleProjectWorkspace(sinon);
 
       assert(properties.hasViewedInstallGuide === false);
-      InstallGuideWebView.tryOpen(properties);
+      // TODO: Change this because this method no longer exists
+      // InstallGuideWebView.tryOpen(properties);
       assert(executeCommand.calledWith('appmap.openWorkspaceOverview'));
       assert(properties.hasViewedInstallGuide);
     });
@@ -53,7 +54,8 @@ describe('Install guide', () => {
       mockSingleProjectWorkspace(sinon);
 
       assert(properties.hasViewedInstallGuide === false);
-      InstallGuideWebView.tryOpen(properties);
+      // TODO: Change this because this method no longer exists
+      // InstallGuideWebView.tryOpen(properties);
       assert(executeCommand.calledWith('appmap.openWorkspaceOverview') === false);
       assert(properties.hasViewedInstallGuide === false);
     });
@@ -65,7 +67,8 @@ describe('Install guide', () => {
       mockSingleProjectWorkspace(sinon);
 
       assert(properties.hasViewedInstallGuide === false);
-      InstallGuideWebView.tryOpen(properties);
+      // TODO: Change this because this method no longer exists
+      // InstallGuideWebView.tryOpen(properties);
 
       assert(executeCommand.calledWith('appmap.openWorkspaceOverview') === shouldOpen);
       assert(properties.hasViewedInstallGuide === shouldOpen);
